put request edit profile

+++ back-end/src/main/java/com/noobs/codecoach/infrastructure/security/authentication/user/accountverification/AccountVerification.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 public class AccountVerification {
     @Id
     @Column(name = "id")
-    private Long profileId;
+    private int profileId;
 
     @Column(name = "verification_code", nullable = false)
     private String verificationCode;
@@ -24,7 +24,9 @@ public class AccountVerification {
 
     }
 
-    public AccountVerification(Long profileId, String verificationCode){
+
+
+    public AccountVerification(int profileId, String verificationCode){
         this.profileId = profileId;
         this.verificationCode = verificationCode;
     }
@@ -33,7 +35,7 @@ public class AccountVerification {
         return verificationCode;
     }
 
-    public Long getId() {
+    public int getId() {
         return profileId;
     }
